[FIX] [WIP] Fixing the code structure, so that 'qbench' can be used as an imported package from an arbitrary location.
