+++ qbench/fourier/__init__.py
@@ -26,8 +26,11 @@ from typing import Union
 
 import numpy as np
 
+import sys
+
 from ._cli import add_fourier_parser
-from ._components import FourierComponents
+# from ._components import FourierComponents
+from ._components.components import FourierComponents
 from ._models import (
     FourierDiscriminationAsyncResult,
     FourierDiscriminationSyncResult,
